Text with autocompletion
Adding a line to get the input element if the current element is a text field with autocompletion. The idea to to move up to the parent and then find the input field which is the original element if it was a text field (without autocompletion) or the input field further downwards.